Change Double to BigDecimal

+++ src/main/java/iit/cloudcopmuting/satelite/dao/SateliteManagementDaoImpl.java
@@ -1,5 +1,6 @@
 package iit.cloudcopmuting.satelite.dao;
 
+import java.math.BigDecimal;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
@@ -41,8 +42,8 @@ public class SateliteManagementDaoImpl implements SateliteManagementDao{
 
 				ps.setString(1, payload.getName());
 				ps.setString(2, payload.getDescription());
-				ps.setDouble(3, payload.getDistance());
-				ps.setDouble(4, payload.getHealth());
+				ps.setBigDecimal(3, payload.getDistance());
+				ps.setBigDecimal(4, payload.getHealth());
 
 				return ps;
 			}
@@ -92,8 +93,8 @@ public class SateliteManagementDaoImpl implements SateliteManagementDao{
 			satelite.setSateliteID((Integer) row.get("sateliteID"));
 			satelite.setName((String) row.get("name"));
 			satelite.setDescription((String) row.get("description"));
-			satelite.setDistance((Double) row.get("distance"));
-			satelite.setHealth((Double) row.get("health"));
+			satelite.setDistance((BigDecimal) row.get("distance"));
+			satelite.setHealth((BigDecimal) row.get("health"));
 
 			satelites.add(satelite);
 
@@ -124,8 +125,8 @@ public class SateliteManagementDaoImpl implements SateliteManagementDao{
 			satelite.setSateliteID((Integer) row.get("sateliteID"));
 			satelite.setName((String) row.get("name"));
 			satelite.setDescription((String) row.get("description"));
-			satelite.setDistance((Double) row.get("distance"));
-			satelite.setHealth((Double) row.get("health"));
+			satelite.setDistance((BigDecimal) row.get("distance"));
+			satelite.setHealth((BigDecimal) row.get("health"));
 
 			satelites.add(satelite);
